# Enable `variant` P0608R3 in C++17, update LLVM, overhaul `variant`/`any`/`optional` tests