--- a/tests/std/tests/P0088R3_variant_msvc/env.lst
+++ b/tests/std/tests/P0088R3_variant_msvc/env.lst
@@ -1,0 +1,4 @@
+# Copyright (c) Microsoft Corporation.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+RUNALL_INCLUDE ..\usual_17_matrix.lst