--- a/tests/std/tests/P0220R1_any/test.cpp
+++ b/tests/std/tests/P0220R1_any/test.cpp
@@ -63,6 +63,7 @@
 #include "count_new.h"
 #include "test_macros.h"
 
+namespace assign::copy {
 template <class LHS, class RHS>
 void test_copy_assign() {
     assert(LHS::count == 0);
@@ -236,6 +237,7 @@ int run_test() {
 
   return 0;
 }
+} // namespace assign::copy
 // -- END: test/std/utilities/any/any.class/any.assign/copy.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.assign/move.pass.cpp
@@ -261,6 +263,7 @@ int run_test() {
 #include "any_helpers.h"
 #include "test_macros.h"
 
+namespace assign::move {
 template <class LHS, class RHS>
 void test_move_assign() {
     assert(LHS::count == 0);
@@ -342,6 +345,7 @@ int run_test() {
 
   return 0;
 }
+} // namespace assign::move
 // -- END: test/std/utilities/any/any.class/any.assign/move.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.assign/value.pass.cpp
@@ -369,6 +373,7 @@ int run_test() {
 #include "count_new.h"
 #include "test_macros.h"
 
+namespace assign::value {
 template <class LHS, class RHS>
 void test_assign_value() {
     assert(LHS::count == 0);
@@ -552,6 +557,7 @@ int run_test() {
 
   return 0;
 }
+} // namespace assign::value
 // -- END: test/std/utilities/any/any.class/any.assign/value.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.cons/copy.pass.cpp
@@ -576,6 +582,7 @@ int run_test() {
 #include "count_new.h"
 #include "test_macros.h"
 
+namespace ctor::copy {
 template <class Type>
 void test_copy_throws() {
 #if !defined(TEST_HAS_NO_EXCEPTIONS)
@@ -653,6 +660,7 @@ int run_test() {
 
   return 0;
 }
+} // namespace ctor::copy
 // -- END: test/std/utilities/any/any.class/any.cons/copy.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.cons/default.pass.cpp
@@ -678,6 +686,7 @@ int run_test() {
 #include "any_helpers.h"
 #include "count_new.h"
 
+namespace ctor::default_ {
 int run_test()
 {
     {
@@ -701,6 +710,7 @@ int run_test()
 
   return 0;
 }
+} // namespace ctor::default_
 // -- END: test/std/utilities/any/any.class/any.cons/default.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.cons/in_place_type.pass.cpp
@@ -735,6 +745,7 @@ int run_test()
 #include "test_macros.h"
 #include "test_convertible.h"
 
+namespace ctor::in_place_type {
 template <class Type>
 void test_in_place_type() {
     // constructing from a small type should perform no allocations.
@@ -896,6 +907,7 @@ int run_test() {
 
   return 0;
 }
+} // namespace ctor::in_place_type
 // -- END: test/std/utilities/any/any.class/any.cons/in_place_type.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.cons/move.pass.cpp
@@ -922,6 +934,7 @@ int run_test() {
 #include "count_new.h"
 #include "test_macros.h"
 
+namespace ctor::move {
 // Moves are always noexcept. The throws_on_move object
 // must be stored dynamically so the pointer is moved and
 // not the stored object.
@@ -997,6 +1010,7 @@ int run_test()
 
   return 0;
 }
+} // namespace ctor::move
 // -- END: test/std/utilities/any/any.class/any.cons/move.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.cons/value.pass.cpp
@@ -1028,6 +1042,7 @@ int run_test()
 #include "count_new.h"
 #include "test_macros.h"
 
+namespace ctor::value {
 template <class Type>
 void test_copy_value_throws()
 {
@@ -1152,6 +1167,7 @@ int run_test() {
 
   return 0;
 }
+} // namespace ctor::value
 // -- END: test/std/utilities/any/any.class/any.cons/value.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.modifiers/emplace.pass.cpp
@@ -1178,6 +1194,7 @@ int run_test() {
 #include "count_new.h"
 #include "test_macros.h"
 
+namespace modifiers::emplace {
 struct Tracked {
     static int count;
     Tracked() { ++count; }
@@ -1439,6 +1456,7 @@ int run_test() {
 
   return 0;
 }
+} // namespace modifiers::emplace
 // -- END: test/std/utilities/any/any.class/any.modifiers/emplace.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.modifiers/reset.pass.cpp
@@ -1462,6 +1480,7 @@ int run_test() {
 #include "test_macros.h"
 #include "any_helpers.h"
 
+namespace modifiers::reset {
 int run_test()
 {
     // empty
@@ -1500,6 +1519,7 @@ int run_test()
 
   return 0;
 }
+} // namespace modifiers::reset
 // -- END: test/std/utilities/any/any.class/any.modifiers/reset.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.modifiers/swap.pass.cpp
@@ -1525,6 +1545,7 @@ int run_test()
 #include "test_macros.h"
 #include "any_helpers.h"
 
+namespace modifiers::swap_ {
 template <class LHS, class RHS>
 void test_swap() {
     assert(LHS::count == 0);
@@ -1632,6 +1653,7 @@ int run_test()
 
   return 0;
 }
+} // namespace modifiers::swap_
 // -- END: test/std/utilities/any/any.class/any.modifiers/swap.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.observers/has_value.pass.cpp
@@ -1655,6 +1677,7 @@ int run_test()
 #include "test_macros.h"
 #include "any_helpers.h"
 
+namespace observers::has_value {
 int run_test()
 {
     {
@@ -1697,6 +1720,7 @@ int run_test()
 
   return 0;
 }
+} // namespace observers::has_value
 // -- END: test/std/utilities/any/any.class/any.observers/has_value.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.class/any.observers/type.pass.cpp
@@ -1722,6 +1746,7 @@ int run_test()
 #include "test_macros.h"
 #include "any_helpers.h"
 
+namespace observers::type {
 int run_test()
 {
     {
@@ -1745,6 +1770,7 @@ int run_test()
 
   return 0;
 }
+} // namespace observers::type
 // -- END: test/std/utilities/any/any.class/any.observers/type.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.nonmembers/any.cast/any_cast_pointer.pass.cpp
@@ -1773,6 +1799,7 @@ int run_test()
 #include "test_macros.h"
 #include "any_helpers.h"
 
+namespace nonmembers::cast::pointer {
 // Test that the operators are properly noexcept.
 void test_cast_is_noexcept() {
     std::any a;
@@ -1928,6 +1955,7 @@ int run_test() {
 
   return 0;
 }
+} // namespace nonmembers::cast::pointer
 // -- END: test/std/utilities/any/any.nonmembers/any.cast/any_cast_pointer.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.nonmembers/any.cast/any_cast_reference.pass.cpp
@@ -1960,6 +1988,7 @@ int run_test() {
 #include "count_new.h"
 #include "test_macros.h"
 
+namespace nonmembers::cast::reference {
 // Test that the operators are NOT marked noexcept.
 void test_cast_is_not_noexcept() {
     std::any a;
@@ -2237,6 +2266,7 @@ int run_test() {
 
   return 0;
 }
+} // namespace nonmembers::cast::reference
 // -- END: test/std/utilities/any/any.nonmembers/any.cast/any_cast_reference.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.nonmembers/make_any.pass.cpp
@@ -2263,6 +2293,7 @@ int run_test() {
 #include "count_new.h"
 #include "test_macros.h"
 
+namespace nonmembers::make_any {
 template <class Type>
 void test_make_any_type() {
     // constructing from a small type should perform no allocations.
@@ -2377,6 +2408,7 @@ int run_test() {
 
   return 0;
 }
+} // namespace nonmembers::make_any
 // -- END: test/std/utilities/any/any.nonmembers/make_any.pass.cpp
 
 // -- BEGIN: test/std/utilities/any/any.nonmembers/swap.pass.cpp
@@ -2401,6 +2433,7 @@ int run_test() {
 
 #include "test_macros.h"
 
+namespace nonmembers::swap_ {
 int run_test()
 {
 
@@ -2420,6 +2453,7 @@ int run_test()
 
   return 0;
 }
+} // namespace nonmembers::swap_
 // -- END: test/std/utilities/any/any.nonmembers/swap.pass.cpp
 // LLVM SOURCES END
 // clang-format on
@@ -3142,7 +3176,7 @@ int main() {
 
     ctor::copy::run_test();
     ctor::default_::run_test();
-    ctor::in_place::run_test();
+    ctor::in_place_type::run_test();
     ctor::move::run_test();
     ctor::value::run_test();
 
@@ -3152,8 +3186,6 @@ int main() {
 
     observers::has_value::run_test();
     observers::type::run_test();
-
-    not_literal::run_test();
 
     nonmembers::cast::pointer::run_test();
     nonmembers::cast::reference::run_test();