@StephanTLavavej 2024-06-14
I'm *speculatively* mirroring this to the MSVC-internal repo - please notify me if any further changes are pushed.